+++ utils.py
@@ -4,9 +4,9 @@ import numpy as np
 from ezc3d import c3d
 
 
-def get_range_q(biorbd_model: biorbd.Model):
+def get_range_q(biorbd_model: biorbd.Model) -> tuple[np.ndarray, np.ndarray]:
     """
-
+    Give the ranges of the model
 
     Parameters
     ----------
@@ -15,58 +15,16 @@ def get_range_q(biorbd_model: biorbd.Model):
 
     Returns
     -------
-    range_q
-        The range of the q for each dof
+    q_range_min, q_range_max: tuple[np.ndarray, np.ndarray]
+        The range min and max of the q for each dof
     """
-    min_ranges = np.zeros((biorbd_model.nbQ()))
-    max_ranges = np.zeros((biorbd_model.nbQ()))
-    for dof_nb in range(biorbd_model.nbDof()):
-        seg_id, dof_id = get_segment_and_dof_id_from_global_dof(biorbd_model, dof_nb)
-        range_min = biorbd_model.segment(seg_id).QRanges()[dof_id].min()
-        range_max = biorbd_model.segment(seg_id).QRanges()[dof_id].max()
-        min_ranges[dof_nb] = range_min
-        max_ranges[dof_nb] = range_max
-    return min_ranges, max_ranges
-
-
-def get_segment_and_dof_id_from_global_dof(biorbd_model: biorbd.Model, global_dof: int):
-    """
-    Allow the users to get the segment id which correspond to a dof of the model and the id of this dof in the segment
-
-    Parameters
-    ----------
-    biorbd_model: biorbd.Model
-        The biorbd model
-    global_dof: int
-        The global id of the dof in the model
-
-    Returns
-    -------
-    seg_id: int
-        The id of the segment which correspond to the dof
-    count_dof: int
-         The dof id in this segment
-    """
-    for j, seg in enumerate(biorbd_model.segments()):
-        complete_seg_name = biorbd_model.nameDof()[global_dof].to_string()  # We get "Segment_Name_DofName"
-        seg_name = complete_seg_name.replace("_RotX", "")  # We remove "_DofName"
-        seg_name = seg_name.replace("_RotY", "")
-        seg_name = seg_name.replace("_RotZ", "")
-        seg_name = seg_name.replace("_TransX", "")
-        seg_name = seg_name.replace("_TransY", "")
-        seg_name = seg_name.replace("_TransZ", "")
-
-        if seg.name().to_string() == seg_name:
-            seg_name_new = seg_name
-            seg_id = j
-
-    dof = biorbd_model.nameDof()[global_dof].to_string().replace(seg_name_new, "")
-    dof = dof.replace("_", "")  # we remove the _ "_DofName"
-    count_dof = 0
-    while biorbd_model.segment(seg_id).nameDof(count_dof).to_string() != dof:
-        count_dof += 1
-
-    return seg_id, count_dof
+    q_range_max = []
+    q_range_min = []
+    for i in range(biorbd_model.nbSegment()):
+        segment = biorbd_model.segment(i)
+        q_range_max += [q_range.max() for q_range in segment.QRanges()]
+        q_range_min += [q_range.min() for q_range in segment.QRanges()]
+    return np.array(q_range_min), np.array(q_range_max)
 
 
 def get_unit_division_factor(c3d_file: ezc3d.c3d) -> int:
